+++ src/main/java/com/example/deliveryplatform/common/exception/customException/BaseException.java
@@ -1,4 +1,4 @@
-package com.example.deliveryplatform.common.exception;
+package com.example.deliveryplatform.common.exception.customException;
 
 import com.example.deliveryplatform.common.exception.code.ErrorCode;
