+++ src/main/java/com/example/deliveryplatform/common/exception/handler/GlobalExceptionHandler.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.example.deliveryplatform.common.exception.BaseException;
+import com.example.deliveryplatform.common.exception.customException.BaseException;
 import com.example.deliveryplatform.common.exception.code.ErrorCode;
 import com.example.deliveryplatform.common.exception.dto.ErrorResponse;
