+++ src/main/java/com/example/deliveryplatform/common/security/CustomUserDetailsServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.deliveryplatform.common.exception.BaseException;
+import com.example.deliveryplatform.common.exception.customException.BaseException;
 import com.example.deliveryplatform.common.exception.code.ErrorCode;
 import com.example.deliveryplatform.domain.user.entity.User;
 import com.example.deliveryplatform.domain.user.repository.UserRepository;
